Fix issue on blocked stderr
We were using a piped stderr when launching a subprogram so that we could
display any error messages if it failed. If the program wrote a sufficient
amount to stderr though, it could block on the call to write (since we
weren't clearing the buffer on the other side).

Fix by piping output to a temporary file instead.