Fix infinite loop when scanning BSS
When scanning BSS sections for a valid interpreter state, there was a
potential for the it to loop indefinitely if the next threadp pointer
pointed back to the current thread or if the next frame pointer pointed
back to the current frame.

Fix this by capping the maximum number of threads/frames at a large value
and returning an error if hit.