Don't require symbols on windows
Support the case where we don't have symbol files for Python on windows,
by scanning the BSS section for the intrepreter.

It doesn't look like it's possible to figure out if something is a pointer
just by checking if it points to valid virtual memory on Windows, so
instead just brute force the search (which doesn't seem to have a large
perfomance penalty).