Add a Timer class for consistent sleep times
We weren't accounting for the length of time it took to collect the sample when
sleeping in between samples. Fix by adding a Timer class that calculates
when we should run to sample at a consistent rate, and only sleeping for the
required amount of time.

Also fixes an issue on windows where the minimum sleep time could be as much
as 15ms.