check interpreter addresses from symbols too
When scanning the BSS table, we check that the interpreter addresses are valid too.
However, we just assumed that the interpreter addresses from the symbol lookup
were valid and returned them immediately if found.

When testing windows 7 with python 3.7, this seemed to cause an issue. Fix by also
validating that we can get a valid interpeter state from the address found in the
symbols, and if not fall back to scanning the BSS section.