memmap binaries instead of reading contents
Since we only need small parts of the binaries (headers/symbols/bss sections),
reading the whole file in is sort of inefficient. Instead, memory map
the files in and use goblin to parse the memmaps. Also fix up how
we are parsing mach binaries to be consistent with elf/pe files.