Fixes for setting namespace
We were trying to change the namespace even when it wasn't necessary,
causing issues like https://github.com/benfred/py-spy/issues/68.

Fix by comparing the target processes namespace to ours by reading the
softlinks appropiately, and also just logging a warning if we failed
to set the namespace if we failed for any reason.