fix regression with process exitted check
We were checking for specific types of OS errors to see if the
process exitted, but with the latest code these could be wrapped
in an remoteprocess::Error. Fix.

Also if the error rate went to 100% the console viewer would freeze,
fix by adding display check to when we're incrementing the error count.