Don't hardcode interp_head offset
For python 3.7, we were hardcoding the offset of the head interpreter
inside the _PyRuntimeState object. This was only valid for 64bit operating
systems, and on 32 bit systems caused warnings to be logged out.

Fix by generating bindings for the PyRuntimeState struct and use the
bindings to figure out the offset of the head interpreter.