Get GIL status on Python 3.7+
Get the GIL status by looking at pyruntime.gilstate.tstate_current on
python v3.7+. This can't be gotten in a cross platform why by using
bindgen, so provide a lookup function to return the offsets instead.
This also removes the previous attempt to generate bindings for the
pyruntime struct.

Also adds support for finding the offsets in generate_bindings.py, and
a basic integration test to make sure that we can detect gil usage
appropiately using all this.