Allow use as a rust library
This change lets py-spy be used as a rust library. This allows you to write
rust code to connect to a python program and take a stack trace from it.

Most of this change is in adding some basic documentation on how to do this,
and adding a src/lib.rs that exports a couple of the core py-spy classes.
Note that we're only exposing a very minimal set of functionality right now,
and leaving things like the consoleviewer/flamegraph code out of this library.

https://github.com/benfred/py-spy/issues/110