[#112] FreeBSD support
FreeBSD is a somewhat popular system among developers and it would be
nice to have it supported.
In order to support it, we have to patch the remoteprocess subcrate a
bit, as well as bump `proc-maps` and `read-process-memory`
crates version.

This change

[remoteprocess crate]

Adds FreeBSD support for

* process and thread locking (via ptrace)
* getting thread registers (via ptrace)
* getting process' executable and cwd (via libprocstat)
* read_process_memory contract has changed in the latest version, so
  OSX / Linux / Windows code has been adjusted accordingly.

The only thing that is left to be implemented is unwinder /
symbolificator. Therefore `py-spy` on FreeBSD does not support native
stack traces, yet.

[py-spy]

* Adds FreeBSD-specific versions of platform-dependent functions.
* Adds Cirrus CI configuration to run tests against FreeBSD
  platform. In order to set CI up, please refer to
  https://cirrus-ci.org/guide/quick-start/