Add support for writing out Speedscope files
This adds support for writing out speedscope files, as well as writing out raw
call stack data. The speedscope code is adapted from rbspy.

This change also refactors the commandline arguments significantly, changing
to take subcommands (```py-spy dump -p 1234```) instead of flags. While this
is a breaking change, I think it is a better long term solution for py-spy
- and it makes more sense for the '-f' flag to be the file format and '-o'
to be the filename (which would have broken commandline compatability anyways).