add option to dump as JSON (+ other fixes)
This adds an option to dump the current stack traces as JSON data to stdout.
This will help out with scripting the output here.

Also fix minor display issues on windows:The background style was wrong after
bringing up the help menu, and the spinner during record was displaying
an invalid unicode symbol.

This also adds a cmdline implementations for both windows/freebsd.