Support for profiling subprocesses
Add support for profiling subprocesses in the top and record options, via
a ```--subprocesses``` flag. In the record option we show the process
tree in the flamegraph, and the top view we will show the # of processes
and the merged activity across processes.

This refactors how we are collecting samples into a 'Sampler' class,
where the sampling is happening on a different thread. For the subprocesess
case we fan out to a new thread for each python subprocess, and also have
a separate thread to monitor for new child processes every 100ms. The
single process case will only have a single extra thread to collect samples.
In both cases we communicate back to the main thread using a mpsc Channel.

This also adds code to list the child processes to the remoteprocess
crate for Linux/FreeBSD/Windows and OSX.