Fixes for not capturing output
There were a couple issues with #222: this broke the duration
flag , and was hard to distinguish commandline output from py-spy
with commandline output from the application being profiled.
This fixes so the duration flag works again, and prefixes messages
showing that they are coming from py-spy. Also fix problem of
not showing message if we were late in sampling with captured
output.