Add parentheses around thread id to prevent diff flamegraph
The use of the --thread option can result in stack samples that look
like "thread <thread_id> <count>". Inferno and Brendan Gregg's perl
script will treat this as a frame function of with name "thread" and two
counts. This results in the creation of differential flamegraph when not
desired. Changing the frame name to "thread (<thread_id>)" fixes this
issue.