Add AlreadyLocked locking strategy
We already had code to sample with or without locking the target process,
but disabling locking also disabled some other functionality like
native tracing and looking up OS thread ids. For use as a library, its
helpful to have a 3rd option which indicates that the process is already
locked - so we shouldn't attempt to lock it, but can try to get native
traces and the OS thread id.