Fix subprocess sampling
In certain cases subprocess sampling was failing. This was related to the process
info map we were sending back with the samples. When a process exitted we
cleaned up the information about the process from the map, but in another thread
was accessing to display the process hierarchy.

Fix by not returning the process info map to the caller, and instead return a
linked list of process information (that is immutable and ref counted) with
each sample.