Count "samples" as the number of recorded stacks (per thread)
This number is more useful for the user (as opposed to "the number of sampling intervals",
which we had until now). It matches the meaning of "samples" in other profiling tools,
e.g flamegraphs, Linux's "perf", and others.

We do not account samples whose recording is skipped: that is, GIL-less samples if --gil is given,
idle samples unless --idle is given.

Closes: #365