Automatically test new python versions
Every night this will check all the supported versions of python on github actions, and file
a PR to expand our test coverage if new versions of python are released.

(Note that this PR purposefully removes an old version of python for testing, this should get
automatically re-added when this first runs).