Specify speedscope units (#401)
* Use Write trait instead of File for writing

This enables tests to pass in a std::io::Cursor type, for validating
outputs. It is more flexible in general (for example, we could output to
stdout/stderr if we wanted to, or even a network socket).

This change enables validation of the output of the Speedscope format in
a future commit.

* Scale the Speedscope output and set units value

This enables speedscope traces to use the correct time units.

* Use sampling rate from config

This means that if the user changes the sampling rate, the traces will
still be correct.