Upgrade goblin to 0.4.1
Encountered this error:

	Error: Failed to parse python binary
	Reason: Malformed entity: Section 106 size (4277144) + addr (120) is out of bounds. Overflowed: false

(installed "python", which is Python 2.7.6, on an ubuntu:14.04 Docker).

I started looking in goblin, and found this commit 84062ff966c98d390
("Fix size check for section headers") which appears to fix this problem,
so I just upgraded to the newest goblin version available (although 0.3.x
also solves the issue).