#431 Handle offsetted Elf segment case + FreeBSD CI
An Elf segment may have a non-zero offset and that may result in
subtraction overflow when we calculating the symbol addresses.

Apart from that, CI workflows are run within GH Actions with FreeBSD
being notable exception. While FreeBSD is not supported on actions,
the conventional workaround is to run builds on macos runner using a
vagrant box.

This change
- Handles the subtraction overflow case by defaulting to zero.
- removes cirrus config and sets up a FreeBSD GH action job.
- Also bumps proc-maps version in attempt to resolve #431