Add `--class-name` param to prepend class names to methods
Regular methods and classmethods supported. No support for static methods.
This is because the frame object doesn't refer to the function, only its
code, which leaves no other option than to try to find `self` or `cls`
in locals. The alternative of using the `gc` module is horrendously slow
even in the same process and also impractical in a remote process.