fix panic with Python 3.13 shim frames (#787)
File/function names in some Py3.13 frames are empty strings.

It appears that the caller (get_stack_trace) handles this condition and
expects the callee (copy_string) to return an error. The callee however
panics by copying zero bytes into a vector and then calling as_ptr on
it, which returns a wild pointer (0x01), which is then passed to
std::slice::from_raw_parts. This is not OK even if the length passed to
from_raw_parts is 0.